Remove setup.cfg file from the repository
OpenBMC CI has migrated from using `pycodestyle` to `flake8` as part of
this[1] commit. Unlike `pycodestyle` , `flake8` does not rely on the
presence of setup.cfg file in their root path as a trigger, but it runs
on all repositories by default. Hence there is no need of having
setup.cfg file, so removing it from the repository.

[1]: https://github.com/openbmc/openbmc-build-scripts/commit/c5ad7ff440cfd94fc025efbd45a3859475b18820

Change-Id: Ia78ebdc6fc76b619b2ce6a5a1acee4481b95a27b
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>